+++ AppCenterDistribute/AppCenterDistributeTests/MSDistributeTests.m
@@ -35,6 +35,7 @@ static NSString *const kMSTestDistributionGroupId = @"DISTRIBUTIONGROUPID";
 static NSString *const kMSTestDownloadedDistributionGroupId = @"DOWNLOADEDDISTRIBUTIONGROUPID";
 static NSString *const kMSDistributeServiceName = @"Distribute";
 static NSString *const kMSUpdateTokenApiPathFormat = @"/apps/%@/update-setup";
+static NSString *const kMSDefaultURLFormat = @"https://fakeurl.com";
 
 // Mocked SFSafariViewController for url validation.
 @interface SFSafariViewControllerMock : UIViewController
@@ -1324,9 +1325,7 @@ static NSURL *sfURL;
 - (void)testOpenURLInAuthenticationSession API_AVAILABLE(ios(11)) {
 
   // If
-  NSURL *fakeURL = [NSURL URLWithString:@"https://fakeurl.com"];
-  id notificationCenterMock = OCMPartialMock([NSNotificationCenter new]);
-  OCMStub([notificationCenterMock defaultCenter]).andReturn(notificationCenterMock);
+  NSURL *fakeURL = [NSURL URLWithString:kMSDefaultURLFormat];
   id appCenterMock = OCMClassMock([MSAppCenter class]);
   OCMStub([appCenterMock sharedInstance]).andReturn(appCenterMock);
   OCMStub([appCenterMock isSdkConfigured]).andReturn(YES);
@@ -1350,15 +1349,8 @@ static NSURL *sfURL;
   XCTAssertNotNil(self.sut.authenticationSession);
   XCTAssert([self.sut.authenticationSession isKindOfClass:[SFAuthenticationSession class]]);
 
-  // When
-  [notificationCenterMock postNotificationName:UIApplicationDidEnterBackgroundNotification object:nil];
-
-  // Then
-  XCTAssertNil(self.sut.authenticationSession);
-
   // Clear
   [appCenterMock stopMocking];
-  [notificationCenterMock stopMocking];
 }
 
 - (void)testCheckForUpdatesDebuggerAttached {
@@ -2599,7 +2591,7 @@ static NSURL *sfURL;
                               appSecret:kMSTestAppSecret
                 transmissionTargetToken:nil
                         fromApplication:YES];
-  NSString *urlPath = [NSString stringWithFormat:kMSUpdateTokenApiPathFormat, kMSTestAppSecret];
+  NSString *urlPath = [NSString stringWithFormat:kMSDefaultURLFormat, kMSTestAppSecret];
   NSURLComponents *components = [NSURLComponents componentsWithString:urlPath];
   [distributeMock openURLInAuthenticationSessionWith:components.URL];
 
@@ -2615,7 +2607,7 @@ static NSURL *sfURL;
 - (void)testOpenURLInAuthenticationSessionFails API_AVAILABLE(ios(11)) {
 
   // If
-  NSURL *fakeURL = [NSURL URLWithString:@"https://fakeurl.com"];
+  NSURL *fakeURL = [NSURL URLWithString:kMSDefaultURLFormat];
   id appCenterMock = OCMClassMock([MSAppCenter class]);
   OCMStub([appCenterMock sharedInstance]).andReturn(appCenterMock);
   OCMStub([appCenterMock isSdkConfigured]).andReturn(YES);
@@ -2642,9 +2634,8 @@ static NSURL *sfURL;
 - (void)testClearAuthenticationSession API_AVAILABLE(ios(11)) {
 
     // If
+    NSURL *fakeURL = [NSURL URLWithString:kMSDefaultURLFormat];
     id authenticationSessionMock = OCMClassMock([SFAuthenticationSession class]);
-    id notificationCenterMock = OCMPartialMock([NSNotificationCenter new]);
-    OCMStub([notificationCenterMock defaultCenter]).andReturn(notificationCenterMock);
     self.sut = [MSDistribute new];
     [self.sut startWithChannelGroup:OCMProtocolMock(@protocol(MSChannelGroupProtocol))
                           appSecret:kMSTestAppSecret
@@ -2653,15 +2644,14 @@ static NSURL *sfURL;
     self.sut.authenticationSession = authenticationSessionMock;
 
     // When
-    [notificationCenterMock postNotificationName:UIApplicationDidEnterBackgroundNotification object:nil];
+    // If session exists we should clear it before start new.
+    [self.sut openURLInAuthenticationSessionWith:fakeURL];
 
     // Then
-    XCTAssertNil([self.sut authenticationSession]);
     OCMVerify([authenticationSessionMock cancel]);
 
     // Clear
     [authenticationSessionMock stopMocking];
-    [notificationCenterMock stopMocking];
 }
 
 @end